Turned mesontest into on internal module.
